requirements.txt
dateparser==1.2.1
fuzzywuzzy==0.18.0
jdatetime==5.2.0
python-telegram-bot==22.0
Requests==2.32.3
